Implement Issue #40

+++ API/app/Http/Controllers/Api/GroceryItemController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\GroceryItem;
+use App\Models\ShoppingList;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 
@@ -20,11 +21,6 @@ class GroceryItemController extends Controller
 
     public function store(Request $request)
     {
-        Log::info("Running store method now for GroceryItemController on " . print_r($request->all(), true));
-
-        foreach ($request->all() as $key => $value) {
-            Log::info("gettype($key) === " . gettype($value));
-        }
 
         try {
             $request->validate([
@@ -36,6 +32,17 @@ class GroceryItemController extends Controller
 
             Log::info("Validation passed for request: " . print_r($request->all(), true));
 
+            // Authorize user
+            $userId = $request->user_id;
+            // This is how we get the user_id of the resource for the grocery item 
+            // (i.e. check its corresponding shopping list)
+            $shoppingList = ShoppingList::findOrFail($request->shopping_list_id);
+            
+            // Sharing will be a feature added here later
+            if (strcmp($shoppingList->user_id, $userId)) {
+                return response()->json(["errorMessage" => "Unauthorized Request"], 403);
+            }
+
             $groceryItem = GroceryItem::create($request->all());
             return response()->json($groceryItem, 201);
         } catch (\Illuminate\Validation\ValidationException $e) {
@@ -45,10 +52,21 @@ class GroceryItemController extends Controller
         }
     }
 
-
-    public function show($id)
+    // The $id is actually the id of the shopping list (it's a foreign key)
+    public function show(Request $request, $id)
     {
         $groceryItems = GroceryItem::where('shopping_list_id', $id)->get();
+        // Authorize user
+        $userId = $request->user_id;
+        // This is how we get the user_id of the resource for the grocery item 
+        // (i.e. check its corresponding shopping list)
+        $shoppingList = ShoppingList::findOrFail($id);
+        
+        // Sharing will be a feature added here later
+        if (strcmp($shoppingList->user_id, $userId)) {
+            return response()->json(["errorMessage" => "Unauthorized Request"], 403);
+        }
+        
         return response()->json($groceryItems, 200);
     }
 
@@ -64,6 +82,18 @@ class GroceryItemController extends Controller
         // Find the grocery item by ID and update the fields
         $groceryItem = GroceryItem::findOrFail($id);
         Log::info("Found groceryItem: " . print_r($groceryItem, true));
+        
+        // Authorize user
+        $userId = $request->user_id;
+        // This is how we get the user_id of the resource for the grocery item 
+        // (i.e. check its corresponding shopping list)
+        $shoppingList = ShoppingList::findOrFail($groceryItem->shopping_list_id);
+        
+        // Sharing will be a feature added here later
+        if (strcmp($shoppingList->user_id, $userId)) {
+            return response()->json(["errorMessage" => "Unauthorized Request"], 403);
+        }
+        
         $groceryItem->name = $validatedData['name'];
         $groceryItem->quantity = $validatedData['quantity'];
         $groceryItem->is_food = $validatedData['is_food'];
@@ -72,10 +102,21 @@ class GroceryItemController extends Controller
         return response()->json($groceryItem, 200);
     }    
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         try {
             $groceryItem = GroceryItem::findOrFail($id);
+            
+            // Authorize user
+            $userId = $request->user_id;
+            // This is how we get the user_id of the resource for the grocery item 
+            // (i.e. check its corresponding shopping list)
+            $shoppingList = ShoppingList::findOrFail($groceryItem->shopping_list_id);
+            
+            // Sharing will be a feature added here later
+            if (strcmp($shoppingList->user_id, $userId)) {
+                return response()->json(["errorMessage" => "Unauthorized Request"], 403);
+            }
             Log::info("Deleting item: " . print_r($groceryItem, true));
             $groceryItem->delete();
